feat: replace previous popup with an opened window
