+++ src/common/CTFlowAI.tsx
@@ -28,7 +28,6 @@ const CTFlowAI = () => {
             <OptionsDropdown />
           </HStack>
         </HStack>
-        {openAIKey ? <TaskUI /> : <SetAPIKey />}
       </Box>
     </ChakraProvider>
   );
