+++ src/pages/Common/styles.css
@@ -125,6 +125,9 @@
 .px-4 {
   padding: 0 1em;
 }
+.pl-4 {
+  padding-left: 1em;
+}
 
 /* Margin Utils */
 .m-4 {
@@ -146,6 +149,9 @@
 .mt-12 {
   margin-top: 3em;
 }
+.ml-4 {
+  margin-left: 1em;
+}
 .mr-1 {
   margin-right: 0.25em;
 }
@@ -154,6 +160,9 @@
 }
 .mr-4 {
   margin-right: 1em;
+}
+.mr-4-i {
+  margin-right: 1em !important;
 }
 .mr-5 {
   margin-right: 1.5em;
@@ -209,4 +218,63 @@
 
 .cypress-trigger:hover {
   background: #e9e9e9;
+}
+
+.flex {
+  display: flex;
+}
+
+.justify-between {
+  justify-content: space-between;
+}
+
+.items-center {
+  align-items: center;
+}
+
+.block {
+  display: block;
+}
+
+.static {
+  position: static;
+}
+
+.fixed {
+  position: fixed;
+}
+
+.bottom-1 {
+  bottom: 1rem;
+}
+
+.right-1 {
+  right: 1rem;
+}
+
+.rounded-md {
+  border-radius: 0.375rem;
+}
+
+button.no-default {
+  background: transparent;
+  border: none;
+  padding: 0;
+  margin: 0;
+  cursor: pointer;
+  outline: none;
+}
+
+button.bg-common {
+  background-color: #080a0b;
+}
+
+button.px-4 {
+  padding-right: 1em;
+  padding-left: 1em;
+}
+
+button.py-2 {
+  padding-top: 0.5em;
+  padding-bottom: 0.5em;
 }
